requirements.txt
annotated-types==0.6.0
asttokens==2.4.1
blinker==1.7.0
Bootstrap-Flask==2.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
executing==2.0.1
Flask==2.3.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
icecream==2.1.3
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.25
sqlmodel==0.0.16
typing_extensions==4.10.0
urllib3==2.2.1
Werkzeug==3.0.0
WTForms==3.0.1
